move touch tracker to diff file
